+++ nisystemlink/clients/alarm/__init__.py
@@ -1,0 +1,3 @@
+from nisystemlink.clients.alarm._alarm_client import AlarmClient
+
+# flake8: noqa